+++ at_client/lib/src/client/remote_secondary.dart
@@ -1,9 +1,12 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:at_client/at_client.dart';
 import 'package:at_client/src/client/secondary.dart';
+import 'package:at_client/src/manager/at_client_manager.dart';
+import 'package:at_client/src/preference/at_client_preference.dart';
+import 'package:at_client/src/util/at_client_util.dart';
 import 'package:at_commons/at_builders.dart';
+import 'package:at_commons/at_commons.dart';
 import 'package:at_lookup/at_lookup.dart';
 import 'package:at_utils/at_utils.dart';
 import 'package:internet_connection_checker/internet_connection_checker.dart';
@@ -47,8 +50,7 @@ class RemoteSecondary implements Secondary {
       verbResult = await executeVerb(builder);
       verbResult = verbResult.replaceFirst('data:', '');
     } on AtClientException catch (e) {
-      logger.severe(
-          'Exception occurred in processing the verb ${e.errorCode} - ${e.errorMessage}');
+      logger.severe('Exception occurred in processing the verb ${e.message}');
     }
     return verbResult;
   }
@@ -99,7 +101,10 @@ class RemoteSecondary implements Secondary {
   }
 
   Future<String?> findSecondaryUrl() async {
-    return await AtLookupImpl.findSecondary(_atSign, _preference.rootDomain, _preference.rootPort);
+    var secondaryAddress = await AtClientManager.getInstance()
+        .secondaryAddressFinder!
+        .findSecondary(_atSign);
+    return secondaryAddress.toString();
   }
 
   Future<bool> isAvailable() async {
@@ -115,13 +120,16 @@ class RemoteSecondary implements Secondary {
       var internetAddress = await InternetAddress.lookup(host);
       //TODO getting first ip for now. explore best solution
       var addressCheckOptions =
-      AddressCheckOptions(internetAddress[0], port: int.parse(port));
-      var addressCheckResult = await InternetConnectionChecker().isHostReachable(addressCheckOptions);
+          AddressCheckOptions(internetAddress[0], port: int.parse(port));
+      var addressCheckResult = await InternetConnectionChecker()
+          .isHostReachable(addressCheckOptions);
       return addressCheckResult.isSuccess;
     } on Exception catch (e) {
-      logger.severe('Secondary server unavailable due to Exception: ${e.toString()}');
+      logger.severe(
+          'Secondary server unavailable due to Exception: ${e.toString()}');
     } on Error catch (e) {
-      logger.severe('Secondary server unavailable due to Error: ${e.toString()}');
+      logger
+          .severe('Secondary server unavailable due to Error: ${e.toString()}');
     }
     return false;
   }